+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `already_linear` flag, for already linearized data.
 - Add support for skipping flat fielding step.
+- Add Gaussian filter for optional noise suppression
 
 ### Fixed
 - Properly apply FDK weighting.
@@ -19,3 +20,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 [Unreleased]: https://github.com/cicwi/SliceRecon/compare/v1.0.0-rc.1...develop
+